Add method for setting property if not defined
Added a StaticUtils.setSystemPropertyIfNotAlreadyDefined method that
can be used to set the value of a system property, but only if that
property is not already set.  If the property already has a value,
then the existing value will be retained regardless of whether it is
the same as or different from the provided value.