Use PropertyManager throughout the LDAP SDK
Update most existing uses of system properties within the LDAP SDK
to use the new PropertyManager class.  This helps simplify the code
(especially in cases where the value needs to be parsed as a Boolean
or numeric value), and makes it possible to use environment
variables to control the behavior as an alternative to system
properties.