Remove ManageCertificatesSecurityManager
Removed the ManageCertificatesSecurityManager class and its
associated test case.  This class was used in an initial
implementation of the manage-certificates tool, which relied on
invoking the Java keytool utility for certain functions, and the
security manager was used to prevent keytool from killing the JVM
by calling System.exit.  manage-certificates was long ago updated so
that it no longer relies on keytool and no longer needs the security
manager, but the security manager was not removed.  It's not part of
the public API, so this doesn't affect backward compatibility in any
way, and security manager functionality has been deprecated from
later versions of Java and may be removed, so there's no good reason
to keep it around.