Further updates to the txn settings control
Made a further change to the transaction settings request control to
add the ability to specify the single-writer lock behavior, which
could previously only be controlled through the server
configuration.  We also removed the ability to specify the scoped
lock timeout, as the backend lock timeout will be used for that
purpose (as well as for the backend exclusive and single-writer
locks).