Add property manager caching support
Updated the PropertyManager class to make it possible to cache
property values for faster access with less contention.  Caching is
disabled by default, but you can enable it by specifying a maximum
cache duration.  Additional methods are provided to clear the cache,
and to populate the cache based on the currently defined system
properties and environment variables.