Add a W3C trace context request control
Added client-side support for a W3C trace context request control
that can be used to convey information for use in distributed tracing
(e.g., via OpenTelemetry).  This control was originally defined in
the ForgeRock Directory Server (now known as PingDS), and will also
be supported in the Ping Identity Directory Server.