Add a discard results search result listener
Added a "discard results" search result listener that can be used in
cases where a search should be performed, but the actual contents of
any applicable search result entries or references are not needed.
The number of entries and references returned will still be
available in the search result.  For example, this may be useful for
testing purposes, if a search should be performed and you want to
verify the result and potentially the number of entries and/or
references returned, but you don't need to examine their contents.